+++ xdis/opcodes/opcode_39.py
@@ -24,6 +24,7 @@ from xdis.opcodes.base import(
     extended_format_RETURN_VALUE,
     finalize_opcodes,
     init_opdata,
+    jabs_op,
     rm_op,
     update_pj3
     )
@@ -54,17 +55,32 @@ rm_op(l, "BEGIN_FINALLY", 53)
 rm_op(l, "WITH_CLEANUP_START", 81)
 rm_op(l, "WITH_CLEANUP_FINISH", 82)
 rm_op(l, "END_FINALLY", 88)
+rm_op(l, "BUILD_LIST_UNPACK", 149)
+rm_op(l, "BUILD_MAP_UNPACK", 150)
+rm_op(l, "BUILD_MAP_UNPACK_WITH_CALL", 151)
+rm_op(l, "BUILD_TUPLE_UNPACK", 152)
+rm_op(l, "BUILD_SET_UNPACK", 153)
+rm_op(l, "BUILD_TUPLE_UNPACK_WITH_CALL", 158)
 rm_op(l, "CALL_FINALLY", 162)
 rm_op(l, "POP_FINALLY", 163)
 
 
 # These are new since Python 3.9
 
-#          OP NAME              OPCODE  POP PUSH
-#-----------------------------------------------
-def_op(l, 'RERAISE',                48,   3, 0)
-def_op(l, 'WITH_EXCEPT_START',      49,   0, 1)
-def_op(l, 'LOAD_ASSERTION_ERROR',   74,   0, 1)
+#          OP NAME               OPCODE  POP PUSH
+#------------------------------------------------
+def_op(l, 'RERAISE',                 48,   3, 0)
+def_op(l, 'WITH_EXCEPT_START',       49,   0, 1)
+def_op(l, 'LOAD_ASSERTION_ERROR',    74,   0, 1)
+def_op(l, 'LIST_TO_TUPLE',           82,   1, 1)
+
+def_op(l, 'IS_OP',                  117,   2, 1)
+jabs_op(l, 'JUMP_IF_NOT_EXC_MATCH', 121,   2, 0)
+def_op(l, 'CONTAINS_OP',            118,   2, 1)
+def_op(l, 'LIST_EXTEND',            162,   2, 1)
+def_op(l, 'SET_UPDATE',             163,   2, 1)
+def_op(l, 'DICT_MERGE',             164,   2, 1)
+def_op(l, 'DICT_UPDATE',            165,   2, 1)
 
 format_value_flags = opcode_38.format_value_flags