+++ NEWS.md
@@ -1,5 +1,11 @@
-5.0.2 2020-06-28
+5.0.3 2020-07-28
 ================
+
+* Add versions 3.8.5, 3.7.8, and 3.6.11
+* Clarify changes to 3.8 ROT_FOUR
+* Update 3.9 magics and opcodes
+
+5.0.2 2020-07-25
 
 * Add Python 3.8.4 as a 3.8 release
 * pydisasm.py Python 3.3 tolerance